[TE] Support TianLong SHCA IB
Add USE_SHCA for TianLong SHCA NICs with extended 17-bit LIDs. Convert
LIDs via shca_17b_types.h helpers and widen lid fields to uint32_t across
Transfer Engine and TENT RDMA paths.

Gate yalantinglibs IBV RPC (YLT_ENABLE_IBV) behind USE_SHCA=OFF so SHCA
builds leave ib_socket untouched and fall back to TCP for store/RPC.

+++ mooncake-transfer-engine/include/transfer_metadata.h
@@ -46,7 +46,7 @@ class TransferMetadata {
    public:
     struct DeviceDesc {
         std::string name;
-        uint16_t lid;
+        uint32_t lid;
         std::string gid;
         std::string eid;  // for ub
 
@@ -157,7 +157,7 @@ class TransferMetadata {
 
     struct HandShakeDesc {
         std::string local_nic_path;
-        uint16_t local_lid = 0;
+        uint32_t local_lid = 0;
         std::string local_gid;
         std::string peer_nic_path;
 #ifdef USE_UB
