+++ mooncake-transfer-engine/include/transport/rdma_transport/rdma_endpoint.h
@@ -177,14 +177,14 @@ class RdmaEndPoint {
 
     std::vector<uint32_t> qpNum() const;
 
-    int doSetupConnection(const std::string &peer_gid, uint16_t peer_lid,
+    int doSetupConnection(const std::string &peer_gid, uint32_t peer_lid,
                           std::vector<uint32_t> peer_qp_num_list,
                           Status connected_status = CONNECTED,
                           std::string *reply_msg = nullptr,
                           SetupConnectionFailureInfo *failure_info = nullptr);
 
     int doSetupConnection(int qp_index, const ibv_gid &peer_gid,
-                          uint16_t peer_lid, uint32_t peer_qp_num,
+                          uint32_t peer_lid, uint32_t peer_qp_num,
                           int local_gid_index, std::string *reply_msg = nullptr,
                           SetupConnectionFailureInfo *failure_info = nullptr);
