+++ mooncake-transfer-engine/include/transport/rdma_transport/rdma_transport.h
@@ -103,7 +103,7 @@ class RdmaTransport : public Transport {
    private:
     int allocateLocalSegmentID();
 
-    int refreshLocalDeviceDesc(const std::string &device_name, uint16_t lid,
+    int refreshLocalDeviceDesc(const std::string &device_name, uint32_t lid,
                                const std::string &gid);
 
     int preTouchMemory(void *addr, size_t length);
