+++ mooncake-transfer-engine/tent/include/tent/transport/rdma/endpoint.h
@@ -166,11 +166,11 @@ class RdmaEndPoint : public std::enable_shared_from_this<RdmaEndPoint> {
     }
 
    private:
-    int setupAllQPs(const std::string& peer_gid, uint16_t peer_lid,
+    int setupAllQPs(const std::string& peer_gid, uint32_t peer_lid,
                     std::vector<uint32_t> peer_qp_num_list,
                     std::string* reply_msg = nullptr);
 
-    int setupOneQP(int qp_index, const std::string& peer_gid, uint16_t peer_lid,
+    int setupOneQP(int qp_index, const std::string& peer_gid, uint32_t peer_lid,
                    uint32_t peer_qp_num, std::string* reply_msg = nullptr);
 
     // Returns the pool segment owning qp_index, or nullptr when no pools are
